refactor: Move buttons to repair window
